Break on expressions after comments
This is intended to insert EOL and spaces for injecting comments.

To be able to do so, we annotate all positions right after comments.

autoBreakNln is changed to preserveBreakNln to make use of information
regarding comments and directives.

There are a few exceptions which should be handled in the future.