[VS] Use standard hookup mechanism
The previous code was using an IMouseProcessorProvider to note the
creation of a new F# based ITextView instance.  The comment indicated
this was known to be incorrect but didn't know a better route.  I
changed the code to use an IWpfTextViewCreationListener instead.  This
is the standard way of listening for IWpfTextView creation