Fix record definition using Stroustrup style (#2484)
* Add stroustrup record definition tests

Tests whether or not record definitions with accessibility modifiers using the stroustrup style are formatted correctly

* Fix private record definition

* Add additional test

Adds test using the internal modifier. Just to be safe.

* Remove redundant test

* Update CHANGELOG.md

* Update CHANGELOG.md

* Adjust stroustrup record definition formatting

Previous formatting would result in a compiler error if there was a trailing comment after the accessibility modifier.

* Remove redundant test.

This is already a tested case and is therefore redundant.